# swagger_server/models/credit_card.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class CreditCard(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, lastdigits: int = None, expiration: str = None, holder_name: str = None):  # noqa: E501
        """CreditCard - a model defined in Swagger

        :param lastdigits: The lastdigits of this CreditCard.  # noqa: E501
        :type lastdigits: int
        :param expiration: The expiration of this CreditCard.  # noqa: E501
        :type expiration: str
        :param holder_name: The holder_name of this CreditCard.  # noqa: E501
        :type holder_name: str
        """
        self.swagger_types = {
            'lastdigits': int,
            'expiration': str,
            'holder_name': str
        }

        self.attribute_map = {
            'lastdigits': 'lastdigits',
            'expiration': 'expiration',
            'holder_name': 'holderName'
        }
        self._lastdigits = lastdigits
        self._expiration = expiration
        self._holder_name = holder_name

    @classmethod
    def from_dict(cls, dikt) -> 'CreditCard':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CreditCard of this CreditCard.  # noqa: E501
        :rtype: CreditCard
        """
        return util.deserialize_model(dikt, cls)

    @property
    def lastdigits(self) -> int:
        """Gets the lastdigits of this CreditCard.

        ostatnie 4 cyfry karty  # noqa: E501

        :return: The lastdigits of this CreditCard.
        :rtype: int
        """
        return self._lastdigits

    @lastdigits.setter
    def lastdigits(self, lastdigits: int):
        """Sets the lastdigits of this CreditCard.

        ostatnie 4 cyfry karty  # noqa: E501

        :param lastdigits: The lastdigits of this CreditCard.
        :type lastdigits: int
        """

        self._lastdigits = lastdigits

    @property
    def expiration(self) -> str:
        """Gets the expiration of this CreditCard.

        data waznosci karty w formacie MM/YY  # noqa: E501

        :return: The expiration of this CreditCard.
        :rtype: str
        """
        return self._expiration

    @expiration.setter
    def expiration(self, expiration: str):
        """Sets the expiration of this CreditCard.

        data waznosci karty w formacie MM/YY  # noqa: E501

        :param expiration: The expiration of this CreditCard.
        :type expiration: str
        """

        self._expiration = expiration

    @property
    def holder_name(self) -> str:
        """Gets the holder_name of this CreditCard.

        Imie i nazwisko właściciela karty  # noqa: E501

        :return: The holder_name of this CreditCard.
        :rtype: str
        """
        return self._holder_name

    @holder_name.setter
    def holder_name(self, holder_name: str):
        """Sets the holder_name of this CreditCard.

        Imie i nazwisko właściciela karty  # noqa: E501

        :param holder_name: The holder_name of this CreditCard.
        :type holder_name: str
        """

        self._holder_name = holder_name