# swagger_server/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: Wypożyczalnia Aut - API
  description: API do obsługi wypożyczalni
  version: "1.0"
servers:
  - url: http://localhost:5000/
...
paths:
  /login/login:
    post:
      tags:
        - login
      summary: Loguje do systemu
      description: Loguje uzytkownka do systemu
      operationId: login_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/login_login_body'
      responses:
        "200":
          description: Token sesji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "400":
          description: Niewlasciwe haslo lub login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.login_controller
  /login/register:
    post:
      tags:
        - login
      summary: Rejestruje uzytkownika do systemu
      description: Rejestruje uzytkownika do systemu
      operationId: login_register_post
      requestBody:
        description: Registration data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterData'
        required: true
      responses:
        "201":
          description: Zarejestrowano! Wyslij teraz /sendtoken po token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "400":
          description: "Wystapil blad! Kod bledu \n1 - mail juz uzyty, \n2 - niewlasciwy\
            \ mail, \n3 - login zajety, \n4 - przeprowadz walidacje jeszcze raz, \n\
            5 - inny blad"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_400'
      x-openapi-router-controller: swagger_server.controllers.login_controller
  /login/sendToken:
    post:
      tags:
        - login
      summary: Wysyla token aktywacyjny
      operationId: login_send_token_post
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
      responses:
        "200":
          description: Token wyslany na maila!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "400":
          description: Poczekaj 2 min! za szybko!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.login_controller
  /login/status:
    get:
      tags:
        - login
      summary: Pobierz status konta
      operationId: login_status_get
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
        "400":
          description: Blad argumentow!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_400_1'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.login_controller
  /login/activate:
    post:
      tags:
        - login
      summary: Aktywacja konta
      description: Aktywuje konto w systemie
      operationId: login_activate_post
      parameters:
        - name: activation_token
          in: query
          description: "Token aktywacyjny z maila, losowy pin 6 znakow "
          required: false
          style: form
          explode: true
          schema:
            type: string
            example: "728340"
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
      responses:
        "201":
          description: Konto aktywowane!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "400":
          description: Niewlasciwy token aktywacyjny!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.login_controller
  /login/logout:
    post:
      tags:
        - login
      summary: Wyloguj z systemu
      operationId: login_logout_post
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
      responses:
        "200":
          description: Wylogowano
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.login_controller
  /login/uploadPhotos:
    post:
      tags:
        - login
      summary: Upload images!
      operationId: login_upload_photos_post
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
        - name: side
          in: query
          description: "F jezeli przod, B jezeli tył"
          required: false
          style: form
          explode: true
          schema:
            type: string
      requestBody:
        content:
          image/*:
            schema:
              type: string
              format: binary
        required: true
      responses:
        "200":
          description: Pomyslnie uploadowano!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "400":
          description: Zdjecie juz zaladowane!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.login_controller
  /user/details:
    get:
      tags:
        - user
      summary: Pobiera informacje o uzytkowniku
      operationId: user_details_get
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
      responses:
        "200":
          description: Pomyslnie pobrano dane uzytkownika
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /user/changepasswd:
    post:
      tags:
        - user
      summary: Zmien haslo
      operationId: user_changepasswd_post
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_changepasswd_body'
      responses:
        "200":
          description: Haslo zmienione
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "400":
          description: "Wystapil blad, kod bledu"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_400_2'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /user/updatelocation:
    post:
      tags:
        - user
      summary: Updates user location
      operationId: user_updatelocation_post
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GPSLocation'
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /user/history:
    get:
      tags:
        - user
      summary: Zwraca historie wypozyczen
      operationId: user_history_get
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
        - name: pagelength
          in: query
          description: Number of records to return
          required: false
          style: form
          explode: true
          schema:
            maximum: 20
            minimum: 1
            type: number
            default: 5
        - name: startindex
          in: query
          description: Start index for paging
          required: false
          style: form
          explode: true
          schema:
            minimum: 0
            type: number
            default: 0
      responses:
        "200":
          description: Lista wypozyczen
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rental'
                x-content-type: application/json
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /user/cards:
    get:
      tags:
        - user
      summary: Lista kart przypisanych do konta
      operationId: user_cards_get
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
      responses:
        "200":
          description: Lista kart
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreditCard'
                x-content-type: application/json
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.user_controller
    post:
      tags:
        - user
      summary: Dodaj karte do konta
      operationId: user_cards_post
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
      requestBody:
        description: Json Application with given schema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_cards_body'
        required: true
      responses:
        "200":
          description: Karta dodana
          content:
            application/json:
              schema:
                type: array
                properties:
                  cardId:
                    type: integer
                    description: ID karty
                    example: 12491824091283
                items:
                  $ref: '#/components/schemas/CreditCard'
                x-content-type: application/json
        "400":
          description: "Wystapil blad, kod bledu"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_400_3'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /user/card/{id}:
    get:
      tags:
        - user
      summary: Info na temat konkretnej karty
      operationId: user_card_id_get
      parameters:
        - name: id
          in: path
          description: Id karty
          required: true
          style: simple
          explode: false
          schema:
            type: integer
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
      responses:
        "200":
          description: Krta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCard'
        "400":
          description: niewlasciwy karty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.user_controller
    delete:
      tags:
        - rent
      summary: Usuwa karte z konta
      operationId: user_card_id_delete
      parameters:
        - name: id
          in: path
          description: Id karty
          required: true
          style: simple
          explode: false
          schema:
            type: integer
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
      responses:
        "200":
          description: Karta usunieta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "400":
          description: niewlasciwe ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.rent_controller
...
components:
  schemas:
    Empty:
      type: object
      properties: { }
      description: Pusty Json
      example: { }
    RegisterData:
      type: object
      properties:
        name:
          type: string
          description: All names of the person
          example: Fryderyk
        surname:
          type: string
          example: Markowski
        plec:
          type: string
          description: M(ale) or F(emale) or O(ther)
          example: M
        login:
          type: string
          example: FrydMark1
        password:
          type: string
          example: Bcrypt(test)
        address:
          type: string
          description: Adres w formacie NrDomuIMieszkania;ulica;kodpocztowy;miasto;kraj
          example: 20;Sienkiewicza;60951;Tarnobrzeg;Polska
        email:
          type: string
          example: FrydMark11@gmail.com
        pesel:
          type: string
          description: Pesel
          example: "60060535351"
      description: Uzywany do przeslania danych do rejestracji
...
    Rental:
      type: object
      properties:
        rentalId:
          type: integer
          description: Id wynajecia
          example: 12521334
        carId:
          type: integer
          description: Id auta
          example: 212419820
        rentalStart:
          type: integer
          description: Czas rozpoczecia wynajecia
          example: 124124123
        rentalEnd:
          type: integer
          description: Czas zakonczenia wynajecia
          example: 123128512084
        rentalCost:
          type: string
          description: Koszt wynajecia
          example: "12.3"
        mileage:
          type: integer
          description: Ile metrów przejechał autem?
          example: 1250812094812
        ended:
          type: boolean
          description: Czy to wynajecie zostalo juz zakonczone?
          example: true
      example:
        rentalCost: "12.3"
        ended: true
        rentalStart: 124124123
        rentalId: 12521334
        carId: 212419820
        rentalEnd: 123128512084
        mileage: 1250812094812
    CreditCard:
      type: object
      properties:
        lastdigits:
          type: integer
          description: ostatnie 4 cyfry karty
          example: 1111
        expiration:
          type: string
          description: data waznosci karty w formacie MM/YY
          example: 10/21
        holderName:
          type: string
          description: Imie i nazwisko właściciela karty
          example: Adam Markowski
      example:
        holderName: Adam Markowski
        expiration: 10/21
        lastdigits: 1111
    User:
      type: object
      properties:
        login:
          type: string
          description: Login uzytkownika
          example: FrydMark1
        email:
          type: string
          description: Email uzytkownika
          example: FrydMark1
        name:
          type: string
          description: Imie
          example: Fryderyk
        surname:
          type: string
          description: Nazwisko
          example: Markowski
        accountType:
          type: string
          description: "PERSONAL - osoba prywatna, COMPANY - firma, ORGANISATION -\
            \ organizacja, UNKNOWN - Inne"
          example: PERSONAL
        status:
          type: string
          description: "ACTIVE - Aktywne konto, INACTIVE - nieaktwyne konto, DOCUMENTS\
            \ - Brak dokumentow, PENDING - wyslano dokumenty, oczekiwanie na potwierdzenie,\
            \ PAYMENT - Brak srodkow na koncie, LOCKED - konto zablokowane, DELETED\
            \ - konto usuniete"
          example: ACTIVE
        balance:
          type: string
          description: stan konta w PLN.GR
          example: "20.30"
      example:
        balance: "20.30"
        surname: Markowski
        accountType: PERSONAL
        name: Fryderyk
        login: FrydMark1
        email: FrydMark1
        status: ACTIVE
...
    GPSLocation:
      type: object
      properties:
        locationLat:
          type: string
          description: Szerokosc geograficzna
          example: "51.312312"
        locationLong:
          type: string
          description: Dlugosc geograficzna
          example: "16.3123123"
    login_login_body:
      type: object
      properties:
        login:
          type: string
          description: Login
          example: TEstasdasd
        password:
          type: string
          description: "Hasło, zahashowane bCryptem"
    inline_response_200:
      type: object
      properties:
        token:
          $ref: '#/components/parameters/SessionToken'
      example:
        token: "{}"
    inline_response_400:
      type: object
      properties:
        error:
          type: integer
          description: "Kod bledu \n1 - mail juz uzyty, \n2 - niewlasciwy mail, \n\
            3 - login zajety, \n4 - przeprowadz walidacje jeszcze raz, \n5 - inny\
            \ blad"
          example: 1
    inline_response_200_1:
      type: object
      properties:
        status:
          type: string
          description: "ACTIVE - Aktywne konto, INACTIVE - nieaktwyne konto, DOCUMENTS\
            \ - Brak dokumentow, PENDING - wyslano dokumenty, oczekiwanie na potwierdzenie,\
            \ PAYMENT - Brak srodkow na koncie, ADMIN - administrator, LOCKED - konto\
            \ zablokowane, DELETED - konto usuniete"
          example: ACTIVE
      example:
        status: ACTIVE
    inline_response_400_1:
      type: object
      properties:
        error:
          type: integer
          description: "Konto nie gotowe. Kody bledu: 1 - konto nieaktwne, 2 - konto\
            \ nie ma dokumentow, 3 - konto jest zablokowane, 4 - konto nie gotowe"
          example: 1
    user_changepasswd_body:
      type: object
      properties:
        newPasswd:
          type: string
          description: "Nowe haslo, zakodowane bcryptem"
        oldPasswd:
          type: string
          description: "Stare haslo, zakodowane bcryptem"
    inline_response_400_2:
      type: object
      properties:
        error:
          type: string
          description: "INVALID - niewlasciwe haslo, USED - haslo w uzyciu, NEWERROR\
            \ - nowe haslo niepoprawne\""
          example: INVALID
    user_cards_body:
      type: object
      properties:
        cardNumber:
          type: integer
          description: CardNumber
          example: 123942385910283
        expirationDate:
          type: string
          description: Data waznosci MM/YY
          example: 10/22
        cardHolder:
          type: string
          description: Wlasciciel karty
          example: Jan Kowalski
        cvv:
          type: integer
          description: CVV karty
          example: 123
        holderAddress:
          type: string
          description: Adress wlasciciela karty
          example: "17 stycznia 24, 10-100 Warszawa, Polska"
    inline_response_400_3:
      type: object
      properties:
        error:
          type: integer
          description: "1 - blad autoryzacji karty, 2 - blad przetwarzania transakcji,\
            \ 3 - inny blad"
          example: 1
...
  parameters:
    SessionToken:
      name: Session_Token
      in: header
      description: Token sesji
      required: true
      style: simple
      explode: false
      schema:
        type: integer
        example: 12943854693802
      example: "1111111111111111"
...